// src/types/api/Food.ts
export enum AisleType {
  FRUITS_AND_VEGETABLES = 'FRUITS_AND_VEGETABLES',
  BUTCHERY = 'BUTCHERY',
  FISHMONGER = 'FISHMONGER',
  DELI = 'DELI',
  DAIRY_PRODUCTS = 'DAIRY_PRODUCTS',
  SAVORY_GROCERY = 'SAVORY_GROCERY',
  SWEET_GROCERY = 'SWEET_GROCERY',
  BAKERY_AND_PASTRY = 'BAKERY_AND_PASTRY',
  FROZEN_PRODUCTS = 'FROZEN_PRODUCTS',
  PREPARED_MEALS = 'PREPARED_MEALS',
  NON_ALCOHOLIC_BEVERAGES = 'NON_ALCOHOLIC_BEVERAGES',
  ALCOHOLIC_BEVERAGES = 'ALCOHOLIC_BEVERAGES',
  CLEANING_PRODUCTS = 'CLEANING_PRODUCTS',
  PERSONAL_CARE = 'PERSONAL_CARE',
  BABY_CARE = 'BABY_CARE',
  PET_SUPPLIES = 'PET_SUPPLIES',
  SEASONAL_PRODUCTS = 'SEASONAL_PRODUCTS',
  HOME_PRODUCTS = 'HOME_PRODUCTS',
}

export interface UpdateFoodApi {
  name?: string;
  aisle?: string;
  icon?: string;
}

export interface CreateFoodApi {
  name: string;
  aisle: string;
}